requirements.txt
appdirs==1.4.4
APScheduler==3.6.3
black==21.4b1
certifi==2020.12.5
click==7.1.2
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.8.1
pycodestyle==2.7.0
pydocstyle==6.0.0
pyflakes==2.3.1
pylama==7.7.1
python-telegram-bot==13.4.1
pytz==2021.1
regex==2021.4.4
six==1.15.0
snowballstemmer==2.1.0
toml==0.10.2
tornado==6.1
typed-ast==1.4.3
typing-extensions==3.7.4.3
tzlocal==2.1
